backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.1.4
yfinance==0.2.36
scikit-learn==1.3.2
python-dotenv==1.0.0
numpy==1.26.3
gunicorn==21.2.0